Move getitem implementation to usm from dpnp
